Refactor: add reduce endpoint folder structure
