+++ packages/sources/reduce/src/adapter.ts
@@ -1,108 +1,21 @@
-import objectPath from 'object-path'
-import { Requester, Validator } from '@chainlink/ea-bootstrap'
-import { util } from '@chainlink/ea-bootstrap'
-import { Execute } from '@chainlink/types'
-import { Decimal } from 'decimal.js'
+import { Builder } from '@chainlink/ea-bootstrap'
+import {
+  Config,
+  ExecuteWithConfig,
+  ExecuteFactory,
+  AdapterRequest,
+  APIEndpoint,
+} from '@chainlink/types'
+import { makeConfig } from './config'
+import * as endpoints from './endpoint'
 
-const inputParams = {
-  reducer: true,
-  initialValue: false,
-  dataPath: false,
-  valuePath: false,
+export const execute: ExecuteWithConfig<Config> = async (request, context, config) => {
+  return Builder.buildSelector(request, context, config, endpoints)
 }
 
-const MAX_DECIMALS = 18
-const DEFAULT_DATA_PATH = 'result'
+export const endpointSelector = (request: AdapterRequest): APIEndpoint =>
+  Builder.selectEndpoint(request, makeConfig(), endpoints)
 
-// Export function to integrate with Chainlink node
-export const execute: Execute = async (request) => {
-  const validator = new Validator(request, inputParams)
-
-  const jobRunID = validator.validated.id
-
-  const { data } = validator.validated
-
-  const dataPath = data.dataPath || DEFAULT_DATA_PATH
-  let inputData = <number[]>objectPath.get(request.data, dataPath)
-
-  // Check if input data is valid
-  if (!inputData || !Array.isArray(inputData) || inputData.length === 0) {
-    throw Error(`Input must be a non-empty array.`)
-  }
-
-  const path = data.valuePath || ''
-  // Check if every input object has a path specified
-  if (path && !inputData.every((val) => objectPath.has(val, path))) {
-    throw Error(`Path '${path}' not present in every item.`)
-  }
-
-  // Get value at specified path
-  const _get = (val: unknown): number => objectPath.get(val, path)
-
-  // Filter undesired values
-  inputData = inputData.filter((val) => !!_get(val))
-
-  // Check if all items are numbers
-  const _isNumber = (val: unknown): boolean => !isNaN(_get(val))
-  if (!inputData.every(_isNumber)) {
-    throw Error(`Not every '${path}' item is a number.`)
-  }
-
-  let result: Decimal
-  switch (data.reducer) {
-    case 'sum': {
-      result = inputData.reduce((acc, val) => {
-        return acc.plus(new Decimal(_get(val)))
-      }, new Decimal(data.initialValue) || new Decimal(0))
-      break
-    }
-    case 'product': {
-      result = inputData.reduce(
-        (acc, val) => acc.mul(new Decimal(_get(val))),
-        new Decimal(data.initialValue) || new Decimal(1),
-      )
-      break
-    }
-    case 'average': {
-      result = inputData.reduce(
-        (acc, val, _, { length }) => acc.plus(new Decimal(_get(val)).div(new Decimal(length))),
-        new Decimal(data.initialValue) || new Decimal(0),
-      )
-      break
-    }
-    case 'median': {
-      const sortedData = inputData.sort((a, b) => _get(a) - _get(b))
-      const mid = Math.ceil(inputData.length / 2)
-      result =
-        inputData.length % 2 === 0
-          ? new Decimal(sortedData[mid]).plus(new Decimal(sortedData[mid - 1])).div(new Decimal(2))
-          : new Decimal(sortedData[mid - 1])
-      break
-    }
-    case 'min': {
-      result = inputData.reduce(
-        (acc, val) => Decimal.min(acc, new Decimal(_get(val))),
-        new Decimal(data.initialValue) || new Decimal(Number.MAX_VALUE),
-      )
-      break
-    }
-    case 'max': {
-      result = inputData.reduce(
-        (acc, val) => Decimal.max(acc, new Decimal(_get(val))),
-        new Decimal(data.initialValue) || new Decimal(Number.MIN_VALUE),
-      )
-      break
-    }
-    default: {
-      throw Error(`Reducer ${data.reducer} not supported.`)
-    }
-  }
-
-  // Avoid printing scientific notation on output with `result.toString()`
-  const resultStr = util.toFixedMax(result, MAX_DECIMALS)
-
-  return Requester.success(jobRunID, {
-    data: { result: resultStr },
-    status: 200,
-  })
+export const makeExecute: ExecuteFactory<Config> = (config) => {
+  return async (request, context) => execute(request, context, config || makeConfig())
 }
