+++ packages/sources/reduce/src/config.ts
@@ -1,0 +1,12 @@
+import { Requester } from '@chainlink/ea-bootstrap'
+import { Config } from '@chainlink/types'
+
+export const NAME = 'REDUCE'
+
+export const DEFAULT_ENDPOINT = 'reduce'
+
+export const makeConfig = (prefix?: string): Config => {
+  const config = Requester.getDefaultConfig(prefix)
+  config.defaultEndpoint = DEFAULT_ENDPOINT
+  return config
+}
